Hanin_CRUD_Interview_Question_Modification

+++ FrontEndAngular/src/app/employees/interviewer/questions-list/questions-list.component.ts
@@ -2,6 +2,8 @@ import { Component, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
 import { QuestionsService } from '../../../Services/questions.service';
 import { questionList } from '../../../Models/questions';
+import { CreateQuestionComponent } from '../create-question/create-question.component';
+import { MatDialog, MatDialogConfig } from '@angular/material/dialog';
 
 @Component({
   selector: 'app-questions-list',
@@ -12,7 +14,7 @@ export class QuestionsListComponent implements OnInit {
  
   questions! : questionList[];
 
-  constructor(private questionsService: QuestionsService , private router: Router ) {
+  constructor(  private dialog :MatDialog, private questionsService: QuestionsService , private router: Router ) {
    }
 
   ngOnInit(): void {
@@ -28,7 +30,18 @@ export class QuestionsListComponent implements OnInit {
       });
   }
 
+  onOpenDialogCreate():void{
+    const dialogConfig = new MatDialogConfig();
+    dialogConfig.disableClose = true;
+    dialogConfig.autoFocus = true;
+    this.dialog.open(CreateQuestionComponent, dialogConfig);
+  }
+
+  
+
   deleteQuestions(id: number) {
+    let confirmation =confirm("Do you really want to delete this question ?")
+    if(confirmation)
     this.questionsService.deleteQuestions(id)
     .subscribe(o =>{        
       this.questions = o;
@@ -41,5 +54,7 @@ questionDetails(id: number){
 updateQuestions(id: number){
   this.router.navigate(['employees/interviewer//updatequestion', id]);
 }
+
+
 
 }
